requirements.txt
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
audioread==3.0.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
ci-sdr==0.0.2
click==8.1.8
ConfigArgParse==1.7
ctc_segmentation==1.7.4
Cython==3.1.0
decorator==5.2.1
diffusers==0.33.1
Distance==0.1.3
editdistance==0.8.1
einops==0.8.1
espnet==202412
espnet-model-zoo==0.1.7
espnet-tts-frontend==0.0.3
fast-bss-eval==0.1.3
filelock==3.18.0
fsspec==2025.3.2
g2p-en==2.1.0
h5py==3.13.0
huggingface-hub==0.31.2
humanfriendly==10.0
hydra-core==1.3.2
HyperPyYAML==1.2.2
idna==3.10
importlib-metadata==4.13.0
inflect==7.5.0
jaconv==0.4.0
jamo==0.4.1
Jinja2==3.1.6
joblib==1.5.0
kaldiio==2.18.1
librosa==0.9.2
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.2.1
nltk==3.9.1
numba==0.60.0
numpy==1.23.5
omegaconf==2.3.0
openai-whisper @ git+https://github.com/openai/whisper.git@dd985ac4b90cafeef8712f2998d62c59c3e62d22
opt_einsum==3.4.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
protobuf==6.31.0
pycparser==2.22
pypinyin==0.44.0
python-dateutil==2.9.0.post0
pytz==2025.2
pyworld==0.3.5
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
resampy==0.4.3
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentencepiece==0.1.97
six==1.17.0
sounddevice==0.5.2
soundfile==0.13.1
speechbrain==1.0.3
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.0
torch-complex==0.4.4
torchaudio==2.7.0
tqdm==4.67.1
transformers==4.51.3
typeguard==4.4.2
typing_extensions==4.13.2
tzdata==2025.2
Unidecode==1.4.0
urllib3==2.4.0
zipp==3.21.0
